Reorganised where start of processing functions are located
